filter_demo/requirements.txt
Flask
openai
python-dotenv
tweepy
Mastodon.py
appwrite